docker_app/requirements.txt
joblib==1.1.0
numpy==1.21.4
Pillow==8.4.0
scikit-learn==1.0.1
scipy==1.7.3
sklearn==0.0
threadpoolctl==3.0.0
